+++ legadilo/feeds/models/feed.py
@@ -20,7 +20,6 @@ from legadilo.reading.models.article import Article, ArticleQuerySet
 from legadilo.reading.models.tag import Tag
 from legadilo.users.models import User
 
-from ...constants import SEARCHED_TEXT_MIN_LENGTH
 from ...types import DeletionResult
 from ...users.models import Notification
 from ...utils.time_utils import utcnow
@@ -232,7 +231,7 @@ class FeedManager(models.Manager["Feed"]):
             .order_by("category__title", "id")
         )
 
-        if len(searched_text) > SEARCHED_TEXT_MIN_LENGTH:
+        if searched_text:
             qs = qs.filter(title__icontains=searched_text)
 
         for feed in qs: