+++ legadilo/feeds/tests/test_views/test_feeds_admin_view.py
@@ -54,7 +54,7 @@ class TestFeedsAdminView:
         }
 
     def test_feed_admin_with_search(self, logged_in_sync_client):
-        response = logged_in_sync_client.get(self.url, {"q": f"<p>{self.feed.title}</p>"})
+        response = logged_in_sync_client.get(self.url, {"q": self.feed.title})
 
         assert response.status_code == HTTPStatus.OK
         assert response.template_name == "feeds/feeds_admin.html"