+++ legadilo/reading/tests/test_views/test_comment_views.py
@@ -10,14 +10,7 @@ from django.urls import reverse
 from legadilo.conftest import assert_redirected_to_login_page
 from legadilo.reading.models import Article, Comment
 from legadilo.reading.tests.factories import CommentFactory
-
-COMMENT = """My comment
-with <em>nasty HTML</em>
-and line breaks
-"""
-CLEANED_COMMENT = """My comment
-with nasty HTML
-and line breaks"""
+from legadilo.utils.testing import extract_htmx_headers
 
 
 @pytest.mark.django_db
@@ -58,14 +51,51 @@ class TestCreateCommentView:
                 self.url,
                 data={
                     "article_id": article.id,
-                    "text": COMMENT,
+                    "text": """My comment
+with <em>nasty HTML</em>
+and line breaks and *Markdown*
+""",
                 },
             )
 
         assert response.status_code == HTTPStatus.OK
         assert article.comments.count() == 1
         comment = article.comments.get()
-        assert comment.text == CLEANED_COMMENT
+        assert (
+            comment.text
+            == """My comment
+with <em>nasty HTML</em>
+and line breaks and *Markdown*"""
+        )
+        assert extract_htmx_headers(response) == {
+            "HX-Reswap": "beforeend",
+            "HX-Retarget": "#all-comments",
+        }
+        assert response.template_name == "reading/partials/comment.html#add-comment-success"
+        assert response.context_data["article_id"] == article.id
+        assert response.context_data["comment_article_form"].errors == {}
+
+    def test_create_empty_because_of_html_tags(self, logged_in_sync_client, user):
+        article = Article.objects.create(user=user)
+
+        response = logged_in_sync_client.post(
+            self.url,
+            data={
+                "article_id": article.id,
+                "text": """<script>alert('My comment')</script>""",
+            },
+        )
+
+        assert response.status_code == HTTPStatus.BAD_REQUEST
+        assert article.comments.count() == 0
+        assert response.context_data["comment_article_form"].errors["text"] == [
+            "Your comment renders to an empty string. Please use correct markdown synthax."
+        ]
+        assert response.context_data["article_id"] == str(article.id)
+        assert extract_htmx_headers(response) == {
+            "HX-Reswap": "innerHTML",
+            "HX-Retarget": "#add-comment-form",
+        }
 
 
 @pytest.mark.django_db