+++ legadilo/reading/views/comment_views.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: 2023-2025 Legadilo contributors
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
+from http import HTTPStatus
 
 from django import forms
 from django.contrib.auth.decorators import login_required
@@ -12,8 +13,8 @@ from django.utils.translation import gettext_lazy as _
 from django.views.decorators.http import require_GET, require_http_methods, require_POST
 
 from legadilo.reading.models import Article, Comment
+from legadilo.reading.templatetags import markdown
 from legadilo.users.user_types import AuthenticatedHttpRequest
-from legadilo.utils.security import full_sanitize
 
 
 class CommentArticleForm(forms.ModelForm):
@@ -34,9 +35,11 @@ class CommentArticleForm(forms.ModelForm):
 
     def clean_text(self):
         text = self.cleaned_data.get("text", "")
-        text = full_sanitize(text)
-        if not text:
-            raise ValidationError("Your text is empty!")
+        rendered_markdown = markdown(text)
+        if not rendered_markdown:
+            raise ValidationError(
+                "Your comment renders to an empty string. Please use correct markdown synthax."
+            )
 
         return text
 
@@ -48,13 +51,29 @@ def create_comment_view(
 ) -> TemplateResponse | HttpResponseBadRequest:
     form = CommentArticleForm(request.POST)
     if not form.is_valid():
-        return HttpResponseBadRequest()
+        return TemplateResponse(
+            request,
+            "reading/partials/comment.html#add-comment-form",
+            {
+                "comment_article_form": form,
+                "article_id": form.data.get("article_id"),
+            },
+            headers={"HX-Retarget": "#add-comment-form", "HX-Reswap": "innerHTML"},
+            status=HTTPStatus.BAD_REQUEST,
+        )
 
     article = get_object_or_404(Article, id=form.cleaned_data["article_id"], user=request.user)
     comment = Comment.objects.create(article=article, text=form.cleaned_data["text"])
 
     return TemplateResponse(
-        request, "reading/partials/comment.html#comment-card", {"comment": comment}
+        request,
+        "reading/partials/comment.html#add-comment-success",
+        {
+            "comment": comment,
+            "comment_article_form": CommentArticleForm(),
+            "article_id": article.id,
+        },
+        headers={"HX-Retarget": "#all-comments", "HX-Reswap": "beforeend"},
     )