+++ legadilo/reading/views/search_views.py
@@ -25,9 +25,7 @@ from legadilo.reading.models.article import (
 )
 from legadilo.types import FormChoices
 from legadilo.users.user_types import AuthenticatedHttpRequest
-from legadilo.utils.security import full_sanitize
 
-from ...constants import SEARCHED_TEXT_MIN_LENGTH
 from ...users.models import User
 from ...utils.validators import is_url_valid
 from .list_of_articles_views import UpdateArticlesForm, update_list_of_articles
@@ -42,9 +40,7 @@ class FeedMultipleChoiceField(forms.ModelMultipleChoiceField):
 
 class SearchForm(forms.Form):
     # Main fields.
-    q = forms.CharField(
-        required=False, min_length=SEARCHED_TEXT_MIN_LENGTH, label=_("Search query")
-    )
+    q = forms.CharField(required=False, min_length=0, label=_("Search query"))
     search_type = forms.ChoiceField(
         required=False,
         choices=constants.ArticleSearchType.choices,
@@ -150,15 +146,7 @@ class SearchForm(forms.Form):
 
     def clean_q(self):
         q = self.cleaned_data["q"]
-        q = full_sanitize(q)
-        q = q.strip()
-        if q and len(q) < self.fields["q"].min_length:  # type: ignore[attr-defined]
-            raise ValidationError(
-                f"You must at least enter {SEARCHED_TEXT_MIN_LENGTH} characters",
-                code="q-too-short-after-cleaning",
-            )
-
-        return q
+        return q.strip()
 
     def clean_search_type(self):
         if not self.cleaned_data.get("search_type"):
@@ -213,15 +201,6 @@ class SearchForm(forms.Form):
             return constants.ReadingListTagOperator.ALL
 
         return constants.ReadingListTagOperator(self.cleaned_data["exclude_tag_operator"])
-
-    def clean_external_tags_to_include(self):
-        cleaned_tags = []
-        for tag in self.cleaned_data.get("external_tags_to_include", []):
-            cleaned_tag = full_sanitize(tag)
-            if cleaned_tag:
-                cleaned_tags.append(cleaned_tag)
-
-        return cleaned_tags
 
     def clean(self):
         super().clean()